fix: manual version bump due to misinterpretation of semantic-release
BREAKING CHANGE: The "Runner" is deprecated. Instead, we switch to the SDK approach. The connery NPM package contains the SDK and CLI required to build plugins. The SDK wraps the plugin into a web server with a standardized API. Now, every plugin is a self-sufficient entity that can be run as a standalone app. Learn more about the new SDK in our docs.